refactor(cms-plugin): split large config files
