+++ libs/cms-plugin/src/components/client/translations-drawer-content.tsx
@@ -1,0 +1,123 @@
+"use client";
+
+import type { CollectionSlug, GlobalSlug, Locale } from "payload";
+
+import { useModal, useTranslation } from "@payloadcms/ui";
+import {
+  formatDrawerSlug,
+  useDrawerDepth,
+} from "@payloadcms/ui/elements/Drawer";
+import { useCallback, useEffect, useState } from "react";
+
+import type {
+  TranslationsKey,
+  TranslationsObject,
+} from "../../translations/types.js";
+import type { AllLocalesText } from "./translations-field-label-types.js";
+
+import styles from "./translations-field-label.module.css";
+import { SelectLocalesDrawer } from "./translations-select-locales-drawer.js";
+import { TranslationsTable } from "./translations-table.js";
+
+type DrawerContentProps = {
+  collectionSlug?: CollectionSlug;
+  currentLocale: Locale;
+  fieldPath: string;
+  globalSlug?: GlobalSlug;
+  id?: string;
+  locales: Locale[];
+  onClose: () => void;
+};
+
+export function DrawerContent({
+  id,
+  collectionSlug,
+  currentLocale,
+  fieldPath,
+  globalSlug,
+  locales,
+  onClose,
+}: DrawerContentProps) {
+  const [data, setData] = useState<AllLocalesText | null>(null);
+  const [isTranslating, setIsTranslating] = useState(false);
+
+  const updateData = useCallback(
+    async function updateData() {
+      const searchParams = new URLSearchParams();
+      if (collectionSlug) {
+        searchParams.set("collection", collectionSlug);
+      }
+      if (globalSlug) {
+        searchParams.set("global", globalSlug);
+      }
+      if (id) {
+        searchParams.set("id", id);
+      }
+      searchParams.set("fieldPath", fieldPath);
+      const result = await fetch(
+        `/api/translations?${searchParams.toString()}`,
+        {
+          credentials: "include",
+        },
+      );
+
+      if (result.ok) {
+        setData(await result.json());
+      } else {
+        setData(null);
+      }
+    },
+    [id, collectionSlug, fieldPath, globalSlug],
+  );
+
+  const { openModal } = useModal();
+  const drawerDepth = useDrawerDepth();
+  const { t } = useTranslation<TranslationsObject, TranslationsKey>();
+
+  useEffect(() => {
+    void (async function () {
+      await updateData();
+    })();
+  }, [updateData]);
+
+  if (!data) {
+    return (
+      <div className={styles.loadingIndicator}>
+        {t("cmsPlugin:common:loading")}
+      </div>
+    );
+  }
+
+  const otherLocales = locales.filter(
+    (locale) => locale.code !== currentLocale.code,
+  );
+  const selectLocalesModalSlug = formatDrawerSlug({
+    slug: `auto-translate-confirmation`,
+    depth: drawerDepth,
+  });
+
+  return (
+    <>
+      <SelectLocalesDrawer
+        collectionSlug={collectionSlug}
+        currentLocale={currentLocale}
+        fieldPath={fieldPath}
+        globalSlug={globalSlug}
+        id={id}
+        isTranslating={isTranslating}
+        locales={locales}
+        modalSlug={selectLocalesModalSlug}
+        setIsTranslating={setIsTranslating}
+        updateData={updateData}
+      />
+      <TranslationsTable
+        currentLocale={currentLocale}
+        data={data}
+        isTranslating={isTranslating}
+        onAutoTranslateClick={() => openModal(selectLocalesModalSlug)}
+        onClose={onClose}
+        otherLocales={otherLocales}
+      />
+    </>
+  );
+}
