+++ libs/cms-plugin/src/components/client/translations-field-label-types.ts
@@ -1,0 +1,5 @@
+import type { SerializedEditorState } from "@payloadcms/richtext-lexical/lexical";
+
+export type AllLocalesText = {
+  value: null | Record<string, SerializedEditorState | string> | undefined;
+};
